extractor, transforms: Some api calls and transforms + repl experiments
I'm stopping here, after having speculatively spec'd in my notes, because I'll
want to pause to learn about reified transactions before I continue on with
validation. When something doesn't want to validate, I won't want to start over,
so it'd be nice to understand how to batch. I think that I need this reified tx
notion to develop that ability.