Extract full pinboard dataset by way of json file
After a lot of instructive fiddling with methods of doing my job using the API,
I hit on the idea of removing some complexity by downloading the "posts/all"
call as a single json file. Examining its structure, I learned that every record
was on its own line. So it was easy to specially handle the first and last
lines, and transform and act on the lines.

Keeping the dataset updated, for as long as that makes sense, can be a job for
the api. My previous tinkerings had worked fine with a date or ten.